Corrected the type in the GetMagickModule define.

+++ MagickCore/log.h
@@ -25,7 +25,7 @@ extern "C" {
 #endif
 
 #if !defined(GetMagickModule)
-# define GetMagickModule()  __FILE__,__func__,(unsigned long) __LINE__
+# define GetMagickModule()  __FILE__,__func__,(const size_t) __LINE__
 #endif
 
 #define MagickLogFilename  "log.xml"
